Pavgame 3.1;
Notable changes:
GameUseCase is changed to public interface that is implemented by both mediator and PavGameViewModel;
Mediator is no longer public class;
PavGameApplication is part of the PavGameViewModel package and only that package has access to the dagger getter.
PavGameApplication no longer has static getter for application, application should be obtained via GetApplication() of appcompatactivity;

Database change:
TO illustrate the good practice of multiple table of relational db, GameEntity will inherit from Pojo super class which will have the uid;
Now, we will set the Uid of the games as the hash created when insert in realtime database via retrofit;
The filed gameDateTime was added, so that when view the games of only one user, you can see the time of the game.
The database was ported, so this is breaking change for the previous version of games;
The game will be inserted in local database only after the confirmation with hash from firebase is received.
In memory data source and remote data source was adapted to ilustrate generified methods in the case we have more than one table in the databse;
PavGameViewHolder is now inner class in GamesAdapter and uses the new viewbinding feature. When getting fro, the local db room (caching db), if id=user then we will consider the datetime field to have the total points, otherwise it will contain the date of a game. This hack was made to not alter too much the structure of the project by manipulating sql queries.

Note that both datasource and repository, as they are public classes with public constructor, can be obtained in the ui, but all their methods are private or protected, so they can't be used in presentation. In domain it can be used only by Repository class, and in data source only by data source class.
On the other hand, mediator inherits from a public interface (as the mvvm with clean architecture states that there should be more use-cases, but it is no longer public class, so the public method won't be accessible without the GameUseCase public interface. The dependency provider will handle the changes, so that the usecase provided to viewmodel will be in fact a mediator object. The viewmodel will override the usecase too, and forward each call to its coresponding method in domain layer (into mediator through the injefcted useCase interface object).

The app at this time uses the last implementation of every plugin lister in gradle scripts and compiles on api 34.
I had to set targetSdk to 33 and compileSdk to 34 to keep the auto testing running, as robolectric still does not support api 34. All tests pases